refactor(text-to-image): moves "/toSharkUIOutput/StabilityAI_Client_Image.ts" to "/toSharkUIOutput/Image/definition.ts"
