+++ src/features/TextToImage/Client/SDXL/toSharkUIOutput/Image/definition.ts
